Cocoa Pod Fixes
Pod issues:
ActiveLabel -> Podspec for this library has not been pushed yet: https://github.com/optonaut/ActiveLabel.swift/issues/197

SwiftIcons -> Library has not been updated to Swift 4. I created a PR for the library and have submitted it for approval. We'll use my fork of the project for now.

XLPagerTabStrip -> Removed and replaced with Tabman.